web-backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-compress
flask-socketio
eventlet
requests
python-dotenv
waitress
